src/main/java/git/w1shm4st3r/recipeapp/converter/NotesCommandToNotes.java
package git.w1shm4st3r.recipeapp.converter;

import git.w1shm4st3r.recipeapp.commands.NotesCommand;
import git.w1shm4st3r.recipeapp.domain.Notes;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class NotesCommandToNotes implements Converter<NotesCommand, Notes> {

    @Synchronized
    @Nullable
    @Override
    public Notes convert(NotesCommand source) {
        if(source == null) {
            return null;
        }
        final Notes notes = new Notes();
        notes.setId(source.getId());
        notes.setRecipeNotes(source.getRecipeNotes());
        return notes;
    }
}
